+++ src/main/java/com/example/final_take_it/model/LoanRequest.java
@@ -15,7 +15,8 @@ public class LoanRequest {
     private Long loanAmount;
 
     @Column(nullable = false)
-    private double loanInterest;
+    private Long loanInterest;
+
 
     @Column(nullable = false)
     private String status;
@@ -63,11 +64,11 @@ public class LoanRequest {
         this.loanAmount = loanAmount;
     }
 
-    public double getLoanInterest() {
+    public Long getLoanInterest() {
         return loanInterest;
     }
 
-    public void setLoanInterest(double loanInterest) {
+    public void setLoanInterest(Long loanInterest) {
         this.loanInterest = loanInterest;
     }
